% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastrand-package.r
\docType{package}
\name{fastrand-package}
\alias{fastrand-package}
\title{fastrand}
\description{
Fast uniform and gaussian (normal) random number generation. The
package can be built to support a threaded backend via OpenMP or Intel
Thread Building Blocks (TBB), as well as a GPU backend via NVIDIA's CUDA.
The package is particularly useful for generating very large vectors for
testing numerical algorithms.
}
\author{
Drew Schmidt
}
\keyword{Package}
