% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runif.r
\name{runif}
\alias{runif}
\title{runif}
\usage{
runif(n, min = 0, max = 1, seed = getseed(), nthreads = getnthreads())
}
\arguments{
\item{n}{The number of values to generate}

\item{min, max}{Parameters for uniform random variables.}

\item{seed}{Seed for the random number generation.}

\item{nthreads}{An optional argument for specifying the number of threads
(in the case of OpenMP or TBB backends) that the routine should (at most)
use.  Does nothing with a CUDA backend.}
}
\description{
Generate from a uniform distribution in parallel.
}
\details{
The function uses thrust's implementation of the minimal standard random
number generation algorithm.
}
\references{
\url{https://thrust.github.io/doc/group__predefined__random.html}
}
