% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thrust.backend.r
\name{thrust.backend}
\alias{thrust.backend}
\title{thrust.backend}
\usage{
thrust.backend()
}
\value{
A string. One of:
\describe{
  \item{omp}{OpenMP (multicore)}
  \item{tbb}{Intel Thread Building Blocks (multicore)}
  \item{cuda}{NVIDIA's CUDA (gpu)}
}
}
\description{
Reports the backend in use.
}
\examples{
fastrand::thrust.backend()

}

